AI_Bot/requirements.txt
PIL.Image == 10.3.0
google.generativeai == 0.5.2